+++ src/core/net/pool.cpp
@@ -1,0 +1,5 @@
+#include "pool.h"
+
+pool::pool(u16 port) {
+    _peer = peer(port);
+}