+++ src/wallet/wallet_main.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
-#include "../core/constants.h"
-#include "../net/peer.h"
+#include <sockpp/tcp_connector.h>
+#include "../core/bytebuff.h"
+#include "../core/crypto/sha/sha256.h"
 
 /**
  * usage: ./wallet --send ADDRESS --amount 56.16 --confirm --push
@@ -18,39 +19,7 @@
  * --list           list of all addresses with balance
  * --push           broadcast transaction to network
  */
-int main(int argc, char** argv){
-    peer<MINER> p;
-    p.connect("tcp://localhost:" NODE_PORT);
-    std::cout << "send to " << p.endpoint() << std::endl;
-    u8 cmd = 0x00;
-    while (cmd < 0xFF) {
-        p.send(vec<u8>{cmd});
-        str data = p.receive();
-        if(data.empty())
-            break;
-        std::cout << "Received " << data << std::endl;
-        cmd += 0x01;
-    }
-    p.disconnect();
-
-    /*
-
-    argparse::ArgumentParser program(argv[0]);
-    program.add_description("Wallet app for fcoin ecosystem");
-
-    program.add_argument("--peer")
-            .help("adds new peer to connect");
-
-    program.add_argument("--peers")
-            .help("adds peers from file");
-
-    try {
-        program.parse_args(argc, argv);
-    }
-    catch (const std::exception& err) {
-        std::cerr << err.what() << std::endl;
-        std::cerr << program;
-        exit(1);
-    }*/
+int main(/*int argc, char **argv*/) {
     return 0;
 }
+